+++ example/lib/main.dart
@@ -4,7 +4,8 @@ import 'package:alice/alice.dart';
 import 'package:alice_example/posts_service.dart';
 import 'package:chopper/chopper.dart';
 import 'package:http/http.dart' as http;
-
+import 'package:alice/core/alice_http_client_extensions.dart';
+import 'package:alice/core/alice_http_extensions.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 
@@ -76,6 +77,31 @@ class _MyAppState extends State<MyApp> {
   void _runHttpRequests() async {
     _runChopperHttpRequests();
     Map<String, dynamic> body = {"title": "foo", "body": "bar", "userId": "1"};
+
+    http
+        .post('https://jsonplaceholder.typicode.com/posts', body: body)
+        .interceptWithAlice(alice, body: body);
+
+    http
+        .get('https://jsonplaceholder.typicode.com/posts')
+        .interceptWithAlice(alice);
+
+    http
+        .put('https://jsonplaceholder.typicode.com/posts/1', body: body)
+        .interceptWithAlice(alice, body: body);
+
+    http
+        .patch('https://jsonplaceholder.typicode.com/posts/1', body: body)
+        .interceptWithAlice(alice, body: body);
+
+    http
+        .delete('https://jsonplaceholder.typicode.com/posts/1')
+        .interceptWithAlice(alice, body: body);
+
+    http
+        .get('https://jsonplaceholder.typicode.com/test/test')
+        .interceptWithAlice(alice);
+
     http
         .post('https://jsonplaceholder.typicode.com/posts', body: body)
         .then((response) {
@@ -109,19 +135,26 @@ class _MyAppState extends State<MyApp> {
     });
 
     httpClient
+        .getUrl(Uri.parse("https://jsonplaceholder.typicode.com/posts"))
+        .interceptWithAlice(alice);
+
+    httpClient
+        .postUrl(Uri.parse("https://jsonplaceholder.typicode.com/posts"))
+        .interceptWithAlice(alice, body: body, headers: Map());
+
+    httpClient
+        .putUrl(Uri.parse("https://jsonplaceholder.typicode.com/posts/1"))
+        .interceptWithAlice(alice, body: body);
+
+    httpClient
+        .getUrl(Uri.parse("https://jsonplaceholder.typicode.com/test/test/"))
+        .interceptWithAlice(alice);
+
+    httpClient
         .postUrl(Uri.parse("https://jsonplaceholder.typicode.com/posts"))
         .then((request) async {
       alice.onHttpClientRequest(request, body: body);
       request.write(body);
-      var httpResponse = await request.close();
-      var responseBody = await utf8.decoder.bind(httpResponse).join();
-      alice.onHttpClientResponse(httpResponse, request, body: responseBody);
-    });
-
-    httpClient
-        .getUrl(Uri.parse("https://jsonplaceholder.typicode.com/posts"))
-        .then((request) async {
-      alice.onHttpClientRequest(request);
       var httpResponse = await request.close();
       var responseBody = await utf8.decoder.bind(httpResponse).join();
       alice.onHttpClientResponse(httpResponse, request, body: responseBody);
@@ -165,7 +198,7 @@ class _MyAppState extends State<MyApp> {
       alice.onHttpClientResponse(httpResponse, request, body: responseBody);
     });
 
-    dio.post("https://jsonplaceholder.typicode.com/posts", data: body);
+    /*dio.post("https://jsonplaceholder.typicode.com/posts", data: body);
     dio.get("https://jsonplaceholder.typicode.com/posts",
         queryParameters: {"test": 1});
     dio.put("https://jsonplaceholder.typicode.com/posts/1", data: body);
@@ -182,7 +215,7 @@ class _MyAppState extends State<MyApp> {
         "https://findicons.com/files/icons/1322/world_of_aqua_5/128/bluetooth.png");
     dio.get(
         "https://upload.wikimedia.org/wikipedia/commons/4/4e/Pleiades_large.jpg");
-    dio.get("http://techslides.com/demos/sample-videos/small.mp4");
+    dio.get("http://techslides.com/demos/sample-videos/small.mp4");*/
   }
 
   void _runHttpInspector() {