+++ lib/core/alice_http_extensions.dart
@@ -1,0 +1,10 @@
+import 'package:alice/alice.dart';
+import 'package:http/http.dart';
+
+extension AliceHttpExtensions on Future<Response> {
+  Future<Response> interceptWithAlice(Alice alice, {dynamic body}) async {
+    Response response = await this;
+    alice.onHttpResponse(response, body: body);
+    return response;
+  }
+}